Reverted the crypto random back to 64-bit build friendly.

+++ kwh-modbus-tests/WindowsFunctions.h
@@ -13,7 +13,7 @@ typedef void *handle;
 class WindowsFunctions
 {
 private:
-	unsigned long _hProv = NULL;
+	unsigned long long _hProv = NULL;
 
 public:
 	WindowsFunctions();
